updates to data for layouts
