+++ components/molecule/AlbumCard.tsx
@@ -1,29 +1,36 @@
-import { FullAlbumDetails } from '@/data/types';
 import Image from 'next/image';
 import React from 'react';
+import { createClient } from '@/utils/supabase/server';
+import { redirect } from 'next/navigation';
+import { FullAlbumDetails } from '@/data/types';
 
-export type Props =
-  Pick<FullAlbumDetails, "artwork_url" | "artist_name" | "title" | "release_year">
-  & { background: boolean, showArtist: boolean }
+export type Props = { albumId: number, background: boolean, showArtist: boolean }
 
 export const id = 'album-card'
 export function propsCheck(p: Props) {
-  return !!p.artwork_url && !!p.title && p.artist_name && p.release_year && p.background && p.showArtist
+  return p && p.albumId
 }
 
-export function AlbumCard(props: Props) {
+type GetAlbumRes = FullAlbumDetails
+
+export async function AlbumCard({ albumId, background = true, showArtist = true }: Props) {
+  const sb = createClient()
+  const { data: album, error: albumError } = await sb.functions.invoke<GetAlbumRes>('get-album', {
+    body: { albumId: albumId }
+  });
+  if (!album || albumError) redirect('/error')
   return (
-    <div className={`flex flex-col mx-auto p-1 rounded-md ${props.background && "bg-gradient-to-br from-gray-700 to-gray-800"}`}>
+    <div className={`flex flex-col mx-auto p-1 rounded-md ${background && "bg-gradient-to-br from-gray-700 to-gray-800"}`}>
       <Image
-        src={props.artwork_url}
-        alt={`album art for ${props.title} - ${props.artist_name}`}
+        src={album.artwork_url}
+        alt={`album art for ${album.title} - ${album.artist_name}`}
         width={250}
         height={250}
         className={'rounded-md'}
       />
-      <em className={'mt-3 ml-1 text-gray-200'}>{props.title}</em>
-      {props.showArtist && <p className={'ml-1 text-gray-200'}>{props.artist_name}</p>}
-      <em className={'ml-1 text-gray-400'}>{props.release_year}</em>
+      <em className={'mt-3 ml-1 text-gray-200'}>{album.title}</em>
+      {showArtist && <p className={'ml-1 text-gray-200'}>{album.artist_name}</p>}
+      <em className={'ml-1 text-gray-400'}>{album.release_year}</em>
     </div>
   )
 }
